Make maven run the unit tests
Moves the unit tests from directory src/main/test to directory src/test/java, to adhere to the default maven directory structure, and removes testSourceDirectory from the pom.xml. In consequence, the unit tests are executed by the maven surefire plugin (whereas, before this change, surefire ran but could not find the tests).

Moreover, adds a small fix to the ZigBeeConverterSwitchOnOffTest, where mockito needs to be told explicitly the type for an 'any' matcher.

Fixes #144
Signed-off-by: Henning Sudbrock <henning.sudbrock@telekom.de>