A channel converter for buttons (e.g., on remote controls or wall switches)
The converter is used for trigger channels of type 'system.button'. The
converter binds to the client cluster implemented by the device with the
button, and emits trigger events when a particular command is received
from the device.

The converter is configured via channel properties that specify on which
commands the converter shall react (a different command can be specified
for short, long, and double button presses, which are the three button
press types by the button system channel).

Since the converter needs to be configured via channel properties,
devices with such buttons need to be specified via a static thing type.