Fixes log statements with format parameters and exceptions
At several places, there are log statements of the form
`logger.info("some format string with one format parameter",
formatParameter, exception)`. This results in not logging the exception
(as it is considered as a format parameter, which is not used in the
format). This commit circumvents this by using `String.format` in those
cases.

Signed-off-by: Henning Sudbrock <henning.sudbrock@telekom.de>