Update device/node config on thing handler initialization
In ZigBeeThingHandler the method 'handleConfigurationUpdate'
(inherited by BaseThingHandler) has been overloaded by a new one which
also accepts a boolean to process also unchanged properties. This is
needed because the 'new' properties are compared with those ones stored
in the thing which are identical during the thing handler
initialization. Hence the 'new' properties would then be ignored.

In ZigBeeThingHandler.doNodeInitialisation() now the method
'handleConfigurationUpdate' is called passing the config properties
stored in the thing. These will overwrite the config properties defined
in zsmartsystems which are sent to the device by default when the ZigBee
network is formed.

In ZigBeeDeviceConfigHandler the method 'updateConfiguration' has been
overloaded by a new one which also accepts a boolean to process also
unchanged properties. This is needed because the 'new' properties are
compared with those ones stored in the thing which are identical during
the thing handler initialization. Hence the 'new' properties would then
be ignored.

Signed-off-by: Tommaso Travaglino <tommaso.travaglino@telekom.de>