Remove deadlock in the reconnect logic of the ZigBeeCoordinatorHandler
The ZigBeeCoordinatorHandler reconnect logic can currently run into a
deadlock as follows:

1. In thread A, the reconnect Job (in `startReconnectJobIfNotRunning`)
calls `initializeZigBee`.
2. The call to `initializeZigBee starts another thread B that
initializes the ZigBee stack.
3. The initialization in thread B fails, resulting in
`networkStateUpdated` being called with new network state `OFFLINE` in
thread B. In consequence, thread B notifies the `reconnectLock` at the
end of the `networkStateUpdated` method.
4. Thread A waits on the reconnectLock at the end of the `run` method
within `startReconnectJobIfNotRunning`.

Because the `notify` call happens before the `wait` call, and the `wait`
call has no timeout, the `wait` call blocks forever.

This commit removes the deadlock by introducing a boolean flag that is
set in addition to notifying the `reconnectLock`, and testing this flag
before starting to wait on the `reconnectLock`.

Signed-off-by: Henning Sudbrock <henning.sudbrock@telekom.de>