Refactoring of some converters to always initialize variables
Some converters used to initialize variables on device initialization in
order to handle attributes updates. After the initilization has been
been split into a device and a converter initialization these variables
were no longer initialized when only the converter was initialized. This
made the initialization run into NPEs. This commit ensures that the
initialization is properly done when only the converter is initialized.

Signed-off-by: Tommaso Travaglino <tommaso.travaglino@telekom.de>