Reduce node discoveries on ZigBee scan
When a ZigBee scan is initiated, the coordinator starts a discovery of all the nodes known by the coordinator.
This can be time consuming and generate lot of traffic, and in presence of many END_DEVICEs could prevent the new
nodes from successfully joining the network.
Skip the discovery if the node has been already discovered.

Signed-off-by: Giuseppe Iannello <giuseppe.iannello@brokenloop.net>