Added firmware upgrade instructions for some Ember
Added instruction for upgrading the firmware for HUSBZB-1 and Telegesis ETRX357USB and perhaps other Ember based controllers using https://github.com/walthowd/husbzb-firmware. I tested these instructions on a HUSBZB-1.

As far as I can tell this tool is Linux only and, even if it did support Windows I don't have a Windows machine to test it on so would not be comfortable writing that section.